pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal_ai_system"
version = "0.1.0"
description = "Personalised Multi Agents"
readme = "README.md"
authors = [
  {name = "Natasha Saldanha", email = "natasha.saldanha67@gmail.com"}
]
maintainers = [
  {name = "Natasha Saldanha", email = "natasha.saldanha67@gmail.com"}
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "rich",
  "streamlit>=1.32.0",
  "google-generativeai>=0.4.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
]
requires-python = ">= 3.12"

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "typecheck" },
]
lint = [
    "ruff",
]
test = [
    "coverage",
    "pytest",
]
typecheck = [
    "ty",
]
docs = [
    "zensical",
    "mkdocstrings-python",
]

[project.urls]
bugs = "https://github.com/Natasha3101/personal_ai_system/issues"
changelog = "https://github.com/Natasha3101/personal_ai_system/releases"
documentation = "https://Natasha3101.github.io/personal_ai_system/"
homepage = "https://github.com/Natasha3101/personal_ai_system"

[project.scripts]
personal_ai_system = "personal_ai_system.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.coverage.run]
branch = true
parallel = true
source = ["src/", "tests/"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_also = [
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@overload",
    "@typing.overload",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true
default-groups = ["dev"]
